完善DBRuleService，，完成addrule，deleterule，clear，但存在参数为对象的问题，需要改为按引用传参。

+++ timeControllerService/DB/RuleModel.cpp
@@ -2,11 +2,11 @@
 namespace DB {
 
 
-    long TimeControllerRule::GetId() {
+    int TimeControllerRule::GetId() {
         return Id;
     }
 
-    void TimeControllerRule::SetId(long id) {
+    void TimeControllerRule::SetId(int id) {
         Id = id;
     }
